Refactored Docker-related functionality and updated version to 1.1.7
